+++ src/ui/mod.rs
@@ -481,9 +481,26 @@ impl<'ui> Ui<'ui> {
     pub fn on_ui_app_event(&mut self, event: UiAppEvent, context: &mut AppContext) -> Result<()> {
         match event {
             UiAppEvent::Modal(modal) => {
-                self.modals.push(modal);
+                if let Some(id) = modal.get_id() {
+                    if let Some(existing) =
+                        self.modals.iter_mut().find(|m| m.get_id().as_ref() == Some(&id))
+                    {
+                        *existing = modal;
+                    } else {
+                        self.modals.push(modal);
+                    }
+                } else {
+                    self.modals.push(modal);
+                }
                 self.on_event(UiEvent::ModalOpened, context)?;
                 context.render()?;
+            }
+            UiAppEvent::PopConfigErrorModal => {
+                if let Some(config_modal) = self.modals.last() {
+                    if config_modal.get_id() == Some("config_error_modal".into()) {
+                        let _ = self.on_ui_app_event(UiAppEvent::PopModal, context);
+                    }
+                }
             }
             UiAppEvent::PopModal => {
                 self.modals.pop();
@@ -657,6 +674,7 @@ impl<'ui> Ui<'ui> {
 pub enum UiAppEvent {
     Modal(Box<dyn Modal + Send + Sync>),
     PopModal,
+    PopConfigErrorModal,
     ChangeTab(TabName),
 }